Senn refactor tests (#149)
* senn: Refactor t-cpp/fcitx

* senn: Refactor t-cpp/fcitx5

* senn: Refactor tests in t/

+++ senn/t-cpp/senn-lib-fcitx-tcp-proxy/run.sh
@@ -1,14 +1,19 @@
 #!/bin/bash -eux
 
-LIB=senn-lib-fcitx--all-systems.a
+LIB=senn-lib-fcitx-kkc-proxy--all-systems.a
+ENGINE=$(cd $(dirname $0)/../ && pwd)/kkc-engine-goto-tokyo.py
+
+if [ ! -f /usr/lib/senn/fcitx/kkc-proxy ]; then
+    exit 1
+fi
 
 rm -f $LIB
 
 ros use ecl
 
 ros \
-    -s senn-lib-fcitx \
-    -e "(asdf:make-build :senn-lib-fcitx :type :static-library :monolithic t :init-name \"init_senn\" :move-here \"./\")" \
+    -s senn-lib-fcitx-kkc-proxy \
+    -e "(asdf:make-build :senn-lib-fcitx-kkc-proxy :type :static-library :monolithic t :init-name \"init_senn\" :move-here \"./\" :prologue-code '(progn (require :sockets) (require :asdf)))" \
     -q
 
 g++ \
@@ -23,7 +28,7 @@ g++ \
     -lecl \
     -o main
 
-DIFF=`./main | grep Sym: | diff - expected.txt`
+DIFF=`./main $ENGINE | grep Sym: | diff - expected.txt`
 if [ $? -ne 0 ]; then
     echo $DIFF
 fi
